Refactor folders * application init, middleware and settings folder app * backends folder for postgres and redis init and create * helpers errors and utils * models * routes * views
