+++ alembic/env.py
@@ -4,8 +4,8 @@ from alembic import context
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
-from models import Base
-from settings import dsn
+from app.settings import dsn
+from models.users import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
