+++ app/__init__.py
@@ -1,0 +1,2 @@
+from app.auth import init_app
+
