+++ views/helpers/params.py
@@ -1,0 +1,7 @@
+default_parameters = [{
+    'in': 'header',
+    'name': 'Authorization',
+    'type': 'string',
+    'format': 'Bearer',
+    'required': 'true'
+}]
